website_flightfrd/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Redis==0.4.0
PyMySQL==1.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
celery==5.3.1
redis==4.6.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==1.10.13
gunicorn==21.2.0
eventlet==0.33.3